feat: restructure packages
