+++ .docker-compose/docker-compose.yml
@@ -15,4 +15,4 @@ services:
     depends_on:
       - redis
     ports:
-      - "5000:5000"
+      - "8080:5000"
