+++ Dockerfile
@@ -10,6 +10,6 @@ COPY . /app
 WORKDIR /app
 
 RUN go mod tidy
-RUN go build -o main .
+RUN go build -o main ./cmd/rest-server
 
 CMD ["/app/main"]
