+++ internal/api/http.go
@@ -1,4 +1,4 @@
-package infra
+package api
 
 import (
 	"github.com/gin-gonic/gin"
