+++ internal/envvar/envs.go
@@ -1,4 +1,4 @@
-package infra
+package envvar
 
 import (
 	"log"
