+++ internal/envvar/envs_test.go
@@ -1,4 +1,4 @@
-package infra
+package envvar
 
 import (
 	"os"
