+++ internal/service/interfaces.go
@@ -1,4 +1,4 @@
-package services
+package service
 
 // ProtocolService Interface to manage the protocol
 type ProtocolService interface {
