+++ internal/service/protocol_cache_service.go
@@ -1,4 +1,4 @@
-package services
+package service
 
 import (
 	"context"
