+++ internal/service/protocol_cache_service_test.go
@@ -1,4 +1,4 @@
-package services
+package service
 
 import (
 	"errors"
