+++ internal/service/protocol_simple_service_test.go
@@ -1,4 +1,4 @@
-package services
+package service
 
 import (
 	"github.com/stretchr/testify/assert"
